Add Getter/Setter for MaxEncoders & Rename bgfx_begin/end to match newer BGFX versions
Added a getter/setter for the MaxEncoders value, which is required if you want to do threaded rendering.

Newer versions of BGFX renamed bgfx_begin/end to bgfx_encoder_begin/end. I'm on a custom build of BGFX so this is causing a conflict for me, so I figure I would push this change to you (which would also require an update to the one hosted here)